Update apt_lazarus.txt
[0] https://securelist.com/operation-applejeus/87553/

```Domains and IPs```. I left IPs due to this phrase: ```We have confirmed that the C2 server addresses (196.38.48[.]121, 185.142.236[.]226) used in this attack have been used by the older variant of Fallchill.``` And they are alive. I think we can do an exception here for these IPs, which currently are not detected by any of used in MT trails\feeds despite on being used by two types of malware: Lazarus and Fallchill.