Update elf_xbash.txt
After the second look, let's add filename-semaphores ```/Xbash```, ```/XbashX``` and ```/XbashY``` as the trail. Not sure for ```/xapache``` and ```/zlibx``` (due to Google results). And ```/libhttpd``` is exactly bad idea to detect as a trail due to Google results. Please, update filename-semaphores list if ```/xapache```, ```/zlibx``` and ```/libhttpd``` are also acceptable to use here.