Create android_buhsam.txt
[0] https://file.gdatasoftware.com/web/de/documents/whitepaper/G_DATA_WhitePaper_-_Analysis_of_Android.Trojan-Spy.Buhsam.A.pdf

P.S. BTW, if this malware is unknown and unnamed, this will be nice sign for ```malicious_php.txt``` trail. But... someone is rather scared of such functionality. ;) And, as I told that in ``````malicious_php.txt``` discussion, I put trails, which **can detect malware and were in use of malware**, and I strongly avoid of putting generic  trails like ```/index.php```, ```/login.php```, ```redirect.php``` and so on.