Create malicious_сpanel.txt
[0] https://viriback.com/30-days-later-97-panels/
[1] https://twitter.com/ViriBack/status/1050543003935416321

Although ```These are admin panels, not C&C addresses... Possibility of a someone to actually be a botnet operator (inside your network): zero to nil.``` connections to malware control panels is also interesting from the point of view of security incidents as an additional semaphore. Especially, if it goes not from IT- or IS-departments IPs.

Ideology is the same of having ```malicious_js.txt``` and ```malicious_php.txt``` trails. in ```suspicious``` folder.

Trails from #298 and #665 goes here.