Update domain.txt
Almost all trails in #1907 use dynamic domains, but I'm not strongly sure, that ```5wya.com``` is dynamic one, because there's no much info about this domain. Due to graphs on VT, this domain leads to CN and HK only. And this domain is marked as one of actors-participants of attack in ```ptsecurity``` survey. So, let him live in ```domain.txt``` trail.