Update ua.txt
``` An interesting point to mention is the User-Agents in use. SpeakUp defines three User-Agents that the infected machine must use in every communication with its C&C. Two of them are MacOS X User-Agents and the third is a hashed string... The unique User-Agents used in the HTTP communication between SpeakUp to the C&C are a possible path to the identity of the threat actor behind this campaign. The unique strings mainly consist of “Mobile/BADDAD“, “Mobile/7B405” and  “E9BC3BD76216AFA560BFB5ACAF5731A3”```

Related to #9397 .