Update darkside.txt
```The FBI reported on May 10, 2021, that IP address 159.65.225.72 is affiliated with the DarkSide ransomware group as a data exfiltration destination. The address is hosted in DigitalOcean, an American cloud provider, in AS14061 within their NYC region. As of the last SecurityScorecard scan on May 5, 2021, the only service running on the C2 IP was SSH (port 22).```