#19323 Use official python base docker image
Use the official python 3 docker image as a base instead of Ubuntu.  This significantly reduced build times as the python image already has a python environment installed.  This also makes the Dockerfile easier to read.

I also removed the `RUN git clone ...` command from the Dockerfile and replaced it with `COPY . .` to prevent double-cloning of the repo.

Since the `docker build` command cannot access parent directories (`COPY .. .` returns an error), I also moved the Dockerfile to the base directory.

Finally, I removed most of the Dockerfile's crontab lines which are hard to read.  These will be replaced with documentation for auto-restarting containers and updating images.